+++ .storybook/storybook.requires.js
@@ -36,8 +36,8 @@ try {
 
 const getStories = () => {
   return {
-    "./src/components/Button/Button.stories.ts": require("../src/components/Button/Button.stories.ts"),
-    "./src/components/Carousel/Carousel.stories.ts": require("../src/components/Carousel/Carousel.stories.ts"),
+    "./src/components/Button/Button.stories.tsx": require("../src/components/Button/Button.stories.tsx"),
+    "./src/components/Carousel/Carousel.stories.tsx": require("../src/components/Carousel/Carousel.stories.tsx"),
     "./src/components/CustomIcon/CustomIcon.stories.tsx": require("../src/components/CustomIcon/CustomIcon.stories.tsx"),
     "./src/assets/stories/Fonts.stories.tsx": require("../src/assets/stories/Fonts.stories.tsx"),
     "./src/assets/stories/Svgs.stories.tsx": require("../src/assets/stories/Svgs.stories.tsx"),
